+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [7.6.0] - Unreleased
+## [7.6.0] - 2021-05-28
 ### Added
+- (js, py) AnnotationBase.RichText and PlainText properties
+- (js, py) Arc.AngleDomain, StartAngle, EndAngle, StartAngleRadians, EndAngleRadians properties
+- (js, py) Arc.ClosestParameter function
+- (js, py) BezierCurve.ToNurbsCurve and Split function
+- (js, py) Surface.SetDomain, GetSpanVector, IsoCurve, GetSurfaceParameterFromNurbsFormParameter, and GetNurbsFormParameterFromSurfaceParameter functions
 - (js, py) Curve.TangentAtStart and TangentAtEnd properties
 - (js, py) Curve.FrameAt, GetCurveParameterFromNurbsFormParameter, and GetNurbsFormParameterFromCurveParameter functions